initial files added copied from other window
Initial files added, copied from other window.

Originally the files wouldn't sync to Git hub, not sure why but I was able to go to Github desktop and press "open in VS code" instead of trying to add files to the cloned repo. Then once this repo was opened in VS code I just copied the initial folders and starting files into the new window which is now synced and showing changes here on Github desktop.